JCBC-925: N1QL helper to build updateFor clause
Motivation:
-----------
When doing an UPDATE, in the set you can provide an "update-for" clause
that looks like FOR variable IN path WHEN condition END. The current DSL
allows to provide that as a single Expression, leaving the user puzzled
about the exact syntax.

Modifications:
--------------
A new category of helpers is introduced in the dsl.clause package. In
this case, a UpdateForClause is provided that allows to guide the user
in constructing the proper update-for Expression.

Added unit tests for the builder in UpdateDslTest.

Result:
-------
User can now be guided by the API in constructing the updateFor clause.

Change-Id: Idf02e0744e32153440e0652c6e5c3b69c86ee5dd
Reviewed-on: http://review.couchbase.org/60922
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>