JCBC-936: Index management only cares about GSI
Motivation:
-----------
The index management API was designed to work only on GSI indexes. It is
important to ensure that a simple name allows unique identification of
an index within this API. Currently it is not made explicit in javadoc
and `listIndexes` can also return VIEW type indexes.

Modifications:
--------------
Javadoc has been improved to mention that the API only ever deals with
GSI type indexes.

ListIndexes has been updated to only consider GSI indexes, and the tests
have been updated accordingly.

Results:
--------
All methods in the Index Management API only deal with GSI, and this has
been made clearer in the documentation.

If in the future a listIndexes variant that lists also VIEW indexes is
needed, we can easily add that.

Change-Id: I9d4e2d43f6296e305881759a16149ecb27130041
Reviewed-on: http://review.couchbase.org/61582
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>