JCBC-938: Multi values in array mutations
(this is still an in-design part of RFC-002, subdoc API)

Motivation:
-----------
RFC-002 introduces the possibility to insert multiple values at once in
an existing array. This is the implementation of this part of the RFC.

Modifications:
--------------
A new internal type, MultiValue, is introduced to facilitate the special
encoding needed for multi-valued array mutations.

FragmentTranscoder now has two abstract methods: one is the classic
encoding path, the other is the MultiValue encoding path.

3 new specs have been added to the subdoc builders: arrayPrependAll,
arrayAppendAll and arrayInsertAll. Each of these takes a Collection of
values. A convenience alternative signature using varargs is also
offered for each (but doesn't offer the createParents parameter).

The different naming (xxxAll suffix) allows to disambiguate some edge
cases relative to varargs for the compiler. It is also close to the java
collections idiom.

Results:
--------
Multi-valued inserts are now possible in the Java subdoc API, using
either a Collection or a vararg enumeration. Tests have been adapted.

Change-Id: I2d3ab0abba64493c12a1d09bb5033155c21d9345
Reviewed-on: http://review.couchbase.org/61712
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Mark Nunberg <mark.nunberg@couchbase.com>