JCBC-940: Rename subdoc builder doXXX methods to execute()
Motivation:
-----------
The original RFC method name for execution of a builder is do(). This is
incompatible eg. with Java, being a keyword. Most other SDKs where this
is also incompatible elected to name the method execute() instead.

Since it is expected that the RFC be amended towards renaming execute(),
the java naming of doMutate() and doLookup() can be harmonized with the
other SDKs.

Modifications:
--------------
doMutate() and doLookup() are now all renamed execute().

Results:
--------
The execution methods of the builders are the same across SDKs where do
cannot be a method name.

Change-Id: Ic8031efa2dd5a0c3280091b4675ea64aea034a15
Reviewed-on: http://review.couchbase.org/62119
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>