JCBC-943: Refactor version information in env
Motivation:
-----------
The various places where the version information can be accessed are
all displaying the information from the core, not the java-client:
 - packageNameAndVersion
 - userAgent
 - toString

Furthermore, there is no way of just getting the version number, or
even the git-generated build number. These two methods have been added
to the core and the same could be done for the client.

Modifications:
--------------
Added clientVersion() and clientBuild() methods to get these specific
version number relative to the java-client.

Modified the construction of SDK_PACKAGE_NAME_AND_VERSION so that it
takes information relative to the java-client, not the core. It also
now includes the git version of the core.

Stopped shadowing USER_AGENT from core and instead defined a specific
static SDK_USER_AGENT. Made sure both SDK_xxx are used as default in the
builder, but the builder still allows to redefine them.
This ensures that the toString method reflects the package name and
version of the java-client.

Added tests around these methods.

Results:
--------
Much better availability of version information in the environment, both
concerning java-client and core-io layers.

Change-Id: I75e7a923864be2d36276bad91a67247bdd7b2a11
Reviewed-on: http://review.couchbase.org/62293
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>