JCBC-948: Only check for error name in prepared retry
Motivation
----------
For prepared statements, part of the current retry process involves
checking for the content of an error message in case of error code 5000.

The message in its entirety is checked, but the message phrasing could
change in the future. There's however an error name in there that is
guaranteed by the query team not to change, so it can be checked instead
of the whole message phrase.

Modifications
-------------
During prepared statement retries, only the "queryport.indexNotFound"
part of the string is now checked.

Result
------
Better future-proofness of the "should we retry that prepared statement"
check.

Change-Id: Id08b8ef78e1e287734ba567f149bfa2b144f4f38
Reviewed-on: http://review.couchbase.org/62358
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>