JCBC-953: Implement MutationState for AT_PLUS querying.
Motivation
----------
This changeset brings the java-client implementation on par with the
RFC on AT_PLUS querying. It implements the MutationState class which
represents an aggregate of one or more MutationTokens and exposes it
as another option on N1qlParams.

Modifications
-------------
The following changes have been made over the previous AT_PLUS
implementation:

  - All the marshalling/exporting of tokens has been moved into the
    MutationState class.
  - The N1qlParams now use the MutationState class marshalling when
    preparing the N1QL query options.
  - Tests have been added for the MutationState class.

Result
------
Compliance with the AT_PLUS RFC and more centralized handling of
many mutation tokens.

Change-Id: I2c6830baed75e17b278e4effbbe0ae54d63322e6
Reviewed-on: http://review.couchbase.org/63621
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>