JCBC-954: Remove Expirmental Annotation from Subdoc APIs
Motivation
----------
Since Couchbase Server 4.5 GA is around the corner and the subdoc
API has been stabilized, the experimental annotations should be
removed to signal the maturity and production grade usage of
the feature.

Modifications
-------------
On all public APIs the stability changed to committed, on some
of the internal APIs it changed to uncommitted since we don't
need to be super strict there.

Result
------
Stable and GA grade subdoc feature - yay!

Change-Id: I658d478049d62cc8bb707bf1b51aeff68f80c862
Reviewed-on: http://review.couchbase.org/63860
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>