JCBC-961: Make isOrderRetained public
Motivation:
-----------
Sometimes it can be useful to be able to programmatically rework the
parameters of a query, notably includeDocs target class (eg. in Spring
Data). But since the isOrderRetained() tuning is not accessible, there's
no way to ensure that it is maintained when resetting the target class.

Modifications:
--------------
ViewQuery.isOrderRetained() is now public.

Results:
--------
Order retained parameter can be kept when modifying the target class of
includeDocs in a ViewQuery.

Change-Id: Ia00d907b62a1b9dde828ce3678c347e4307ce0e5
Reviewed-on: http://review.couchbase.org/64278
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Simon Baslé <simon@couchbase.com>