JCBC-963: Facet name is kept in SearchQuery
Motivation:
-----------
In RFC10 (couchbaselabs/sdk-rfcs#17), some SDKs can directly map the
object representation to JSON. In the previous design of facets, having
the name attribute as part of the SearchFacet made that problematic.
It has been decided that the name of a facet should only be set on the
SearchQuery in which the facet is added (which would also allow reuse of
facets across queries, with different names).

Modifications:
--------------
SearchFacet has no name attribute anymore, moved into the SearchQuery.

Change-Id: I47e5265f663218cc58eb80b1021614229c0d0189
Reviewed-on: http://review.couchbase.org/64390
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Simon Baslé <simon@couchbase.com>