JCBC-957: Add support for consistency in FTS
Motivation
----------
In Couchbase 4.5, FTS queries can have two levels of consistency:
 - unbounded (default)
 - at-plus

The later needs the query to include a vector with vbucketId / sequence
numbers of the earliest acceptable mutations to consider.

Modifications
-------------
Added parameter builder for consistency to SearchQuery:
 - scanConsistency to set a simple consistency via enum (for now just
 NOT_BOUNDED)
 - consistentWith(...) with either documents, fragments or a
 MutationState to set an AT_PLUS consistency.

Added MutationState.exportForFts() export method to get the FTS-desired
format for the vector. For now, only the SearchQuery's index is
supported in the vector.

Result
------
Basic AT_PLUS consistency as well as unbounded consistency can be set
on a SearchQuery.

Change-Id: I8a0e3dd553294b7928bdb6a115f4d5ffc40a4556
Reviewed-on: http://review.couchbase.org/64402
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>