JCBC-968: DocumentFragment does not carry CAS on multi lookup
Motivation
----------
It has been found that under multi lookup scenario, the DocumentFragment
always returns a cas of 0.

Modifications
-------------
Adapt to the related core-io change (JVMCBC-337) and pass on the CAS
value. Note: since in the previous implementation the core response
was not in scope anymore the code has been refactored slightly as
a nested observable to get it in scope.

Result
------
On multi lookup, the CAS is now part of the DocumentFragment.

Change-Id: I6cddd130e1ad86e6ff8c68935bfb1097718c31fb
Reviewed-on: http://review.couchbase.org/64590
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>