JCBC-964: Expose "disableMetrics" as N1qlParam option.
Motivation
----------
If the metrics are not needed as part of the response (to save bytes
on the network) it should be possible to disable them.

Modifications
-------------
This changeset adds the appropriate query param option and adds a test
case to assert that.

Result
------
The metrics can be disabled as part of the response.

Change-Id: I84afb91234bdae391069bb84c95adfd269b96f7c
Reviewed-on: http://review.couchbase.org/64592
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>