JCBC-965: Add option for generic N1QLParams.
Motivation
----------
There are a bunch of undocumented and experimental params
that can be used with the N1QL query engine. It doesn't make
sense to expose them as stable API but it should be possible
to supply them if needed.

Modifications
-------------
This changeset adds a raw API where generic params (but they
still need to be valid json, so the code checks) can be stored
and they are passed down at query time.

Result
------
It is possible to supply generic, raw N1QL params if needed.

Change-Id: I5e141d49c75bdf9d4cbd6707003f8e5d65ee95a9
Reviewed-on: http://review.couchbase.org/64642
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>