JCBC-971: Support Vararg for Subdoc LookupIn
Motivation
----------
Instead of chaining gets together like .get("foo").get("bar") it
should be possible to also shortcut to .get("foo", "bar") while
still supporing the other style.

Modifications
-------------
The exists and get argument on the LookupInBuilder (incl. Async)
now support the String as a vararg.

Result
------
More flexible ways to specify the fields to lookup or check for
existence.

Change-Id: I6a22ae77b9a7814792a0c33ce2ce70d13b8e0b59
Reviewed-on: http://review.couchbase.org/64710
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>