JCBC-973: Fix detection of disabled flush on bucket.
Motivation
----------
A changed introduce in order to fix JVMCBC-199 regressed the
functionality to detect a bucket where flush is disabled but
nonetheless executed.

Modifications
-------------
The code now checks on the broader != success instead of a
specific error code. Note that success for http includes all
kinds of 2XX error codes, as a result even if the server responds
with 201 or 202 it works.

Result
------
FlushDisabledException raised again if flush disabled.

Change-Id: Iff09885637039fcb850710d865a2480c7f1a6575
Reviewed-on: http://review.couchbase.org/64847
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>