JCBC-978: Fix Index DSL "WITH nodes" option
Motivation:
-----------
When creating a N1QL index, the WITH nodes option should be passed to
the service as a JSON array of node names. However, the index DSL is
not exporting an array.

Modifications:
--------------
Added "withXXX" methods that support passing in multiple node names, and
made all nodes-related processing produce a JSON array.

Results:
--------
Indexes can be created on specific node(s) using the Index DSL.

Change-Id: Ifec2784b6eea7dfe07ce68e9d3e7a7489ec20b6a
Reviewed-on: http://review.couchbase.org/64958
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>