JCBC-984: Don't broadcast query plan to non-query nodes.
Motivation
----------
When a query plan needs to be broadcasted to all the nodes
it is important that in an MDS setup those broadcasts are
not sent to non-query nodes since their requests will never
be able to complete (retried forever and time out eventually).

Modifications
-------------
The code has been adapted to filter() out any nodes that do
not have the QUERY service enabled.

Result
------
Query plan is only broadcasted to query nodes.

Change-Id: I9ff92cc28f27135d0bb0586a0df92f29e66fe7a4
Reviewed-on: http://review.couchbase.org/65530
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>