JCBC-987: Add observe support for subdoc mutateIn
Motivation
----------
Since a subdoc mutation is very similar to a regular kv-based
mutation it is important to also optionally check for the given
durability requirements. The support is all there, it just needs
to be exposed properly on both the blocking and non-blocking APIs.

Modifications
-------------
This changeset adds more variants to the execute() method on the
MutateInBuilder (both sync and async) which, after a successful
subdoc mutation, performs the durability requirement polling.

It behaves exactly like the fulldocument counterpart.

Result
------
It is now possible to specify durability requirements for subdoc
mutations.

Change-Id: I7cb65f1983f78d5154a4f5ed016b230208959b16
Reviewed-on: http://review.couchbase.org/66034
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>