JCBC-982: BucketSettings raw and customSettings
Motivation
----------
Methods of BucketSettings don't cover the full range of possible values
either from the server (when getting bucket info) or when configuring
a bucket. It also varies with each major Couchbase version.

Getting the buckets info could be enriched with a generic raw() method,
while configuring a bucket could take additional generic map of custom
settings.

Modifications
-------------
BucketSettings now has a raw() method to represent the raw information
as returned by the server.
It is set when calling build() on the default builder.

BucketSettings for configuration (from client to server) can now have
additional settings through the customSettings() method. These are set
on the builder via the withSetting and withSettings methods.

The insert/update bucket payload is built using a testable method that
will consider the customSettings(), BUT will ignore them if they would
overwrite a setting that has a corresponding method.

Result
------
Bucket settings can now be read fully from the server and also set
fully, even if native methods are missing. This is more future-proof.

Change-Id: I7a9e6a433a177b617168e714bd85d62d07ad47be
Reviewed-on: http://review.couchbase.org/66218
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>