JCBC-992: DNS SRV workaround when IPv6 nameserver
Motivation:
-----------
The Bootstrap class, used to bootstrap the SDK from a DNS SRV record,
relies on the default host machine DNS resolution.
In the case where the DNS server list configured in the host machine
include IPv6 entries, the whole initialization breaks and there's no
way of using a different resolution process.

Modifications:
--------------
An additional overload for Bootstrap is added that accepts a nameserver
IP (should be an IPv4), to use with the "dns://" prefix for direct name
resolution instead of attempting to compile a list of name servers from
the OS.

The Bootstrap test has been amended to use this feature and rely on
Google's public DNS ("8.8.8.8").

Results:
--------
Bootstrap test pass even on a machine where IPv6 name servers are
configured.
Such OS configurations can be worked around by providing an explicit
name server address, if needed.

Change-Id: I690088c090aa11279235c625ed1e81f846d13b1b
Reviewed-on: http://review.couchbase.org/66290
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>