JCBC-901: Introduce Authenticator for implicit credentials
Motivation:
-----------
RFC 007 (couchbaselabs/sdk-rfcs#13) introduces the Authenticator
interface to deal with credentials that are not provided explicitly on
each authenticated method call but centralized in one place.

Modifications:
--------------
Added the Authenticator interface and its first implementation based on
login/password pairs, PasswordAuthenticator.

Internally, Authenticator deals with Credential objects. The contexts in
which Authenticator can provide credential are enumerated in the
CredentialContext enum.

The Authenticator is set on the Cluster only. Buckets (and N1QL/FTS
queries initiated from Buckets) still have a strong dependency to a
single password through the SASL dance and the way connections are
established. However it is possible to manually set credentials for
multiple buckets on a N1qlQuery.

Generally the authenticator should be prepared and set before opening
buckets, once. However, it can be changed later on via  the
`Cluster#authenticate` method. If a reference is kept to it, it can
also be mutated.

The following Cluster methods make use of the Authenticator when no
explicit credential is provided: clusterManager() and
openBucket(String bucketName). The later will also default to an empty
password if no credentials can be found in the Authenticator.

Results:
--------
Implicit credentials can be retrieved from a configurable Authenticator,
which paves the way for further authentication modes support and cluster
level authentication.

Joins on protected buckets now work, provided their credentials have
been set on the N1qlParams object.

Change-Id: Ia58107462cbeb8b3b2b86402a8e5e11c2ee92c3e
Reviewed-on: http://review.couchbase.org/66801
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Simon Baslé <simon@couchbase.com>