JCBC-851: Cluster Level Querying
Motivation:
-----------
Breaking from the historical model of data manipulation at the bucket
level, N1QL allows to perform JOINS over multiple buckets.

Even though it *is* technically possible to perform such a query from
a Bucket (by making sure that all relevant credentials other than the
ones for the root bucket are provided in the N1qlQuery), it is a bit
counterintuitive.

Cluster-level querying aims at providing a more intuitive way of doing
such bucket-spanning queries.

Note that the configuration and resource management of the SDK makes it
so that performing a N1QL query still has the prerequisite that a
bucket is currently opened in the Cluster.

Modifications:
--------------
Mirrored the query API from Bucket/AsyncBucket to Cluster/AsyncCluster.

The sync version delegates to the AsyncCluster, while the async version
attempts to find an open AsyncBucket and delegate the execution of the
query to it. If no open bucket is found, the method will error with an
UnsupportedOperationException.

The query is enriched with the credentials taken from the currently
configured Authenticator. If none are available, the query method will
error with an IllegalStateException.

The function that allows to go from an AsyncN1qlQueryResult to a sync
one has been extracted in N1qlQueryExecutor. It is reused by the query
implementation in CouchbaseCluster.

Results:
--------
N1QL queries spanning multiple buckets (including password protected
ones) can be more intuitively performed from the Cluster/AsyncCluster.

This functionality and its specific error condition (no open bucket)
are tested in an integration test.

Change-Id: I4001c492a7829f713fe6bc98945a19102d9e9da1
Reviewed-on: http://review.couchbase.org/67630
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>