JCBC-1023 Ignore FTS tests if search service not found
Motivation
----------
Since FTS is an individual service, we may not be
enabling it always while running integration tests.

Changes
-------
Add ignoreIfSearchServiceNotFound to test context, which tries a search
query and does not continue with the tests if a
ServiceNotAvailableException is returned.

Results
-------
Fts tests will be ignored if search service is not found

Change-Id: If76b8fa77b09d3597358c659ec5bfe26fa3e6e9d
Reviewed-on: http://review.couchbase.org/70046
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>