JCBC-1021: Print raw data if document decoding failed.
Motivation
----------
To better debug flag and decoding related issues it makes sense to
warn-log all known details at the point of decoding.

Modifications
-------------
Adds a WARN log to the decoding section of the abstract transcoder
which logs and dumps all known information for that specific document
at the time of failure.

Result
------
More insight into why a document failed to decode properly.

Change-Id: I356602c1bb58de03cc281c7e5b3e37ea99cb522b
Reviewed-on: http://review.couchbase.org/70146
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>