JCBC-1024: Add "sort" option to FTS query.
Motivation
----------
Couchbase Server 4.6 introduces the possibility to
apply a custom sorting order (not only on the score in
descending order), the Java SDK needs to expose this
option as well.

Modifications
-------------
A "sort" option is provided which allows the user to specify
an array of sort fields, including the special ones like
"_score" and "_id" as well as using a "-" prefix for descending
order.

Result
------
FTS compatibility for Server 4.6 which allows the user to specify
a custom sort order.

Change-Id: Icbb12bed5cb808b80398da23f8b8735df2c97820
Reviewed-on: http://review.couchbase.org/70148
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>