Revision of the Exceptions constructors with no params
Motivation
----------
All the java client exceptions having a a method with no parameters dont make a super call.
The oracle compiler in compilation time add this call but all the exceptions have as father
 the CouchbaseException and his no params constructor nor have the super call.

This is a problem because the stack trace is never initialized.

Modifications
-------------
Add super() to all the constructors.

Result
------
The Exception is correctly initialized.

Change-Id: I24460e247cebc8d7b4275ec55f22cc7591693485
Reviewed-on: http://review.couchbase.org/71182
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-on: http://review.couchbase.org/71966
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>