JCBC-1037: Fail FTS Query with IndexDoesNotExistException
Motivation
----------
Right now if an index does not exist we don't fail with the right
error message for FTS.

Modifications
-------------
This changeset makes the response explicitly check if the index
does not exist and if so raises an IndexDoesNotExistException
as part of the error chain with a FTS response.

By the way, this also fixes the longer standing issue that the
FTS queries should be skipped if the index is not found, leading
to failing tests.

Result
------
Proper awareness if an FTS index doesn't exist, and also fixing
failing tests and instead ignoring them.

Change-Id: I721677bfbb489ebd92ee9b91ab7325fc2f71cf59
Reviewed-on: http://review.couchbase.org/71856
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>