JCBC-999: Redispatch config messages if failed.
Motivation
----------
If messages going down the config service in core-io, because the
sockets are opened dynamically and other components depend on this
fail fast mode (while not ideal), such messages need to be retried
aggressively to overcome issues where a node is down for example.

Modifications
-------------
Simple retry logic has been added to the config service messages,
retrying after 100ms and as a result providing a better experience,
even if a node is down or not reachable.

In the future, if needed, we can make this configurable but for
now lets just stick with sane defaults.

Result
------
Bucket and Cluster-level config messages have a much higher chance
of succeeding, even under individual node failures.

Change-Id: If27355d444fc411dd933a65e74c07816cb5dd376
Reviewed-on: http://review.couchbase.org/71897
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-on: http://review.couchbase.org/71967