JCBC-1045: Pre-check connection string to avoid IOOB exception
Motivation
----------
A community member reported an IndexOutOfBounds when working with
the cluster manager. Looking at the code it looks like the connection
string did not contain a single host.

Modifications
-------------
The code now checks if there is at least a hostname loadable from
the parsed connection string and if not fails with a helpful
message instead of IOOB.

Result
------
Better error diagnostics when the connection string is not correct.

Change-Id: I40fd2ab89eec30cfe4cc3ac5292daf1e4cc46af3
Reviewed-on: http://review.couchbase.org/74518
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>