JCBC-1043: Add toString for N1QL Rows and Sync Result
Motivation
----------
If SELECT RAW was used to print out a row, the code
would fail since it would not be parsable into a
JsonObject.

Modifications
-------------
The code has been fixed to just turn it into a string
and dump it that way, enabling all kinds of output
instead of just a JsonObject.

This change also adds a toString to the sync N1qlResult
since it makes sense to add it for utility purposes.

Result
------
Propery printing for Rows with Select RAW as well as
sync results.

Change-Id: Ib4103a42fb93fce9b83b82184dee6cacab4c3e82
Reviewed-on: http://review.couchbase.org/74519
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>