JCBC-1062: Optimize LegacyTranscoder encoding for strings
Motivation
----------
The LegacyTranscoder still uses the slower JVM-based encoding for
strings and it rather should use the same optimized encoding
path which we use for RawJsonDocument and others.

Modifications
-------------
This code adds a "fastpath" for the string encoding and moves
all the other options into a nested else block so that the buffer
is only allocated if needed (and not twice w/ overridden by the
optimized method).

Result
------
Faster encoding for the string path.

Change-Id: Id67e7f7b905491b3bb5be5f1bf0552a3fd264b79
Reviewed-on: http://review.couchbase.org/75774
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>